+++ app/src/index.ts
@@ -1,3 +1,4 @@
 import TimepickerUI from './timepicker';
+import { OptionTypes } from './timepicker/utils/types';
 
-export { TimepickerUI };
+export { TimepickerUI, OptionTypes };